method syntax dont need to be curried anymore
